LPS-57312 Match new pattern
