+++ modules/apps/application-list/application-list-taglib/src/META-INF/resources/panel/init.jsp
@@ -14,11 +14,4 @@
  */
 --%>
 
-<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
-
-<%@ taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
-taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
-
-<%@ page import="com.liferay.application.list.taglib.servlet.taglib.display.context.logic.PanelAppContentHelper" %>
-
-<liferay-theme:defineObjects />
+<%@ include file="/init.jsp" %>
