+++ modules/apps/application-list/application-list-taglib/src/META-INF/resources/panel_app/init.jsp
@@ -14,13 +14,4 @@
  */
 --%>
 
-<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
-
-<%@ taglib uri="http://liferay.com/tld/application-list/ui" prefix="application-list-ui" %><%@
-taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
-taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
-
-<%@ page import="com.liferay.application.list.PanelCategory" %><%@
-page import="com.liferay.application.list.taglib.constants.ApplicationListWebKeys" %>
-
-<liferay-theme:defineObjects />
+<%@ include file="/init.jsp" %>
