+++ modules/apps/application-list/application-list-taglib/src/META-INF/resources/panel_category_content/init.jsp
@@ -14,16 +14,4 @@
  */
 --%>
 
-<%@ include file="/taglib/ui/panel_category_content/init.jsp" %>
-
-<%
-PanelCategory panelCategory = (PanelCategory)request.getAttribute("application-list-ui:panel-category-content:panelCategory");
-
-request.setAttribute(ApplicationListWebKeys.PANEL_CATEGORY, panelCategory);
-
-boolean include = panelCategory.include(request, response);
-%>
-
-<c:if test="<%= !include %>">
-	<application-list-ui:panel panelCategory="<%= panelCategory %>" />
-</c:if>
+<%@ include file="/init.jsp" %>
