--- modules/apps/application-list/application-list-taglib/src/META-INF/resources/taglib/ui/panel_category/init.jsp
@@ -1,31 +1,0 @@
-<%--
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
---%>
-
-<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
-
-<%@ taglib uri="http://liferay.com/tld/application-list/ui" prefix="application-list-ui" %><%@
-taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %><%@
-taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
-
-<%@ page import="com.liferay.application.list.PanelApp" %><%@
-page import="com.liferay.application.list.PanelAppRegistry" %><%@
-page import="com.liferay.application.list.PanelCategory" %><%@
-page import="com.liferay.application.list.taglib.constants.ApplicationListWebKeys" %><%@
-page import="com.liferay.application.list.taglib.servlet.taglib.display.context.logic.PanelCategoryHelper" %><%@
-page import="com.liferay.portal.kernel.util.StringPool" %><%@
-page import="com.liferay.portal.kernel.util.StringUtil" %>
-
-<liferay-theme:defineObjects />
