+++ modules/apps/application-list/application-list-taglib/src/com/liferay/application/list/taglib/servlet/taglib/PanelCategoryContentTag.java
@@ -12,16 +12,16 @@
  * details.
  */
 
-package com.liferay.application.list.taglib.servlet.taglib.ui;
+package com.liferay.application.list.taglib.servlet.taglib;
 
 import com.liferay.application.list.PanelCategory;
 
 import javax.servlet.http.HttpServletRequest;
 
 /**
- * @author Adolfo Pérez
+ * @author Eudaldo Alonso
  */
-public class PanelCategoryTag extends BasePanelTag {
+public class PanelCategoryContentTag extends BasePanelTag {
 
 	public void setPanelCategory(PanelCategory panelCategory) {
 		_panelCategory = panelCategory;
@@ -36,16 +36,15 @@ public class PanelCategoryTag extends BasePanelTag {
 
 	@Override
 	protected String getPage() {
-		return _PAGE;
+		return "/panel_category_content/page.jsp";
 	}
 
 	@Override
 	protected void setAttributes(HttpServletRequest request) {
 		request.setAttribute(
-			"application-list-ui:panel-category:panelCategory", _panelCategory);
+			"application-list-ui:panel-category-content:panelCategory",
+			_panelCategory);
 	}
-
-	private static final String _PAGE = "/taglib/ui/panel_category/page.jsp";
 
 	private PanelCategory _panelCategory;
