+++ modules/apps/product-menu/product-menu-web/bnd.bnd
@@ -8,5 +8,5 @@ Include-Resource:\
 	classes,\
 	META-INF/resources=src/META-INF/resources
 Require-Capability:\
-	osgi.extender;osgi.extender="jsp.taglib";uri="http://liferay.com/tld/application-list/ui";version:Version="1.0.0"
+	osgi.extender;osgi.extender="jsp.taglib";uri="http://liferay.com/tld/application-list";version:Version="1.0.0"
 Web-ContextPath: /product-menu-web
