+++ modules/apps/product-menu/product-menu-web/src/META-INF/resources/layout/view/user_personal_panel.jsp
@@ -18,7 +18,7 @@
 
 <c:choose>
 	<c:when test="<%= themeDisplay.isStatePopUp() %>">
-		<application-list-ui:panel-content portletId="<%= themeDisplay.getPpid() %>" />
+		<liferay-application-list:panel-content portletId="<%= themeDisplay.getPpid() %>" />
 	</c:when>
 	<c:otherwise>
 		<aui:container cssClass="panel-manage-frontpage">
@@ -27,7 +27,7 @@
 					<liferay-portlet:runtime portletName="<%= ProductMenuPortletKeys.PRODUCT_MENU %>" />
 				</aui:col>
 				<aui:col width="<%= 75 %>">
-					<application-list-ui:panel-content portletId="<%= themeDisplay.getPpid() %>" />
+					<liferay-application-list:panel-content portletId="<%= themeDisplay.getPpid() %>" />
 				</aui:col>
 			</aui:row>
 		</aui:container>
