+++ modules/apps/product-menu/product-menu-web/src/META-INF/resources/portlet/view.jsp
@@ -21,4 +21,4 @@ PanelCategoryRegistry panelCategoryRegistry = (PanelCategoryRegistry)request.get
 PanelCategory panelCategory = panelCategoryRegistry.getPanelCategory(PanelCategoryKeys.MY_SPACE);
 %>
 
-<application-list-ui:panel-category-content panelCategory="<%= panelCategory %>" />
+<liferay-application-list:panel-category-content panelCategory="<%= panelCategory %>" />
